# perf: optimize NLP processing and remove redundant keyword matching