--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "creative-thinking",
-  "version": "0.6.0",
+  "version": "0.6.1",
   "description": "Three-layer MCP server for structured creative problem-solving techniques",
   "license": "GPL-3.0",
   "author": "@uddhav",