… CURRENT_WORK_PLAN.md
… NLP_OPTIMIZATION.md
… benchmarks/PERFORMANCE_BASELINES.md
… dist/layers/discovery/ProblemAnalyzer.d.ts
… dist/layers/discovery/ProblemAnalyzer.d.ts.map
… dist/layers/discovery/ProblemAnalyzer.js
… dist/layers/discovery/ProblemAnalyzer.js.map
… src/__tests__/temporal-work-design.test.ts